Move fixation and image trial to pic task.
